Expand E2E test coverage and add type coercion support
- Add template rules with data tables to DataTableE2ETest
- Add navigation assignment test with MutableCustomer fixture
- Add record navigation test with PersonRecord getter methods
- Add test for assigning to outputs in when clause
- Extract numeric helpers to shared NumericHelpers runtime class
- Rename Calculator.addMultiply to scaleByThenAdd for clarity
- Refactor SimpleRuleE2ETest to use compile-once-instantiate-many pattern
- Fix navigation assignment to generate setter calls
- Allow navigation assignments to fact properties in validator

+++ iskibal-integration-tests/src/test/java/work/spell/iskibal/e2e/Customer.java
@@ -1,25 +1,24 @@
 package work.spell.iskibal.e2e;
 
-import java.math.BigDecimal;
-
 /**
- * Test fixture class representing a customer for rule tests.
+ * Test fixture class representing a customer for rule tests. Uses int for age
+ * to test type coercion between int and BigDecimal.
  */
 public class Customer {
 
 	private final String name;
-	private final BigDecimal age;
+	private final int age;
 
 	public Customer(String name, int age) {
 		this.name = name;
-		this.age = BigDecimal.valueOf(age);
+		this.age = age;
 	}
 
 	public String getName() {
 		return name;
 	}
 
-	public BigDecimal getAge() {
+	public int getAge() {
 		return age;
 	}
 }
